Refactor serialization code
